Issue
If a user booked a date in my hotel and after that he/she want to cancel the booking is it possible to do this functionality with jet booking plugin? If yes, then How and how will I able to charge a percent of that whole money from the booking money he/she paid?
Please let me know if it is possible or not. 

please add this kind of functionality